+++ app/Interfaces/NewsRepositoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+namespace App\Interfaces;
+
+use Classes\News;
+interface NewsRepositoryInterface
+{
+    public function listAll();
+    public function add(News $news);
+    public function delete(int $id);
+}